Add new LocalPayment encodable model
